Fix RemoveDuplicateIndexDataTest

+++ app/bundles/InstallBundle/Tests/InstallFixtures/ORM/RemoveDuplicateIndexDataTest.php
@@ -32,6 +32,9 @@ class RemoveDuplicateIndexDataTest extends MauticMysqlTestCase
      */
     private $fixture;
 
+    private $hadAssetsXrefTableBeforeTest = false;
+    private $hadListXrefTableBeforeTest   = false;
+
     protected function setUp(): void
     {
         parent::setUp();
@@ -40,10 +43,19 @@ class RemoveDuplicateIndexDataTest extends MauticMysqlTestCase
         $this->fixture->setContainer($this->getContainerFake());
     }
 
+    /**
+     * We only want to drop the email_assets_xref and email_list_xref tables
+     * if we created them in this test, otherwise they should stay for other tests.
+     * This prevents errors like "Table 'mautictest.email_list_xref' doesn't exist" in CI.
+     */
     protected function tearDown(): void
     {
-        $this->dropTable('email_assets_xref');
-        $this->dropTable('email_list_xref');
+        if ($this->hadAssetsXrefTableBeforeTest) {
+            $this->dropTable('email_assets_xref');
+        }
+        if ($this->hadListXrefTableBeforeTest) {
+            $this->dropTable('email_list_xref');
+        }
 
         parent::tearDown();
     }
@@ -82,7 +94,7 @@ class RemoveDuplicateIndexDataTest extends MauticMysqlTestCase
 
     private function createTables(): void
     {
-        $this->connection->exec('
+        $this->hadAssetsXrefTableBeforeTest = ($this->connection->exec('
             CREATE TABLE IF NOT EXISTS email_assets_xref
             (
                 email_id int unsigned not null,
@@ -91,9 +103,9 @@ class RemoveDuplicateIndexDataTest extends MauticMysqlTestCase
                 INDEX IDX_asset_id (asset_id),
                 INDEX IDX_email_id (email_id)
             )
-        ');
+        ') > 0) ? false : true;
 
-        $this->connection->exec('
+        $this->hadListXrefTableBeforeTest = ($this->connection->exec('
             CREATE TABLE IF NOT EXISTS email_list_xref
             (
                 email_id int unsigned not null,
@@ -102,7 +114,7 @@ class RemoveDuplicateIndexDataTest extends MauticMysqlTestCase
                 INDEX IDX_email_id (email_id),
                 INDEX IDX_leadlist_id (leadlist_id)
             )
-        ');
+        ') > 0) ? false : true;
     }
 
     private function hasTableIndexForColumn(string $table, string $column): bool
